requirements.txt
python-telegram-bot==20.6
openai
pygame
gTTS
selenium
requests
inputimeout
python-dotenv